Add support for the prf extension
The WebAuthn Level 3 draft spec is unclear what parts of extension processing are for inputs and what are for outputs.

It also says that it uses the hmac-secret FIDO-CTAP extension when talking to the authenticator, but the input it describes doesn't match the inputs described by hmac-secret, and in one case the input relies on information that comes from the authenticator, which makes no sense.